chore: restructure project and cleanup (#80)
* chore(devcontainer): restructure and update devcontainer configuration files to separate form other concerns

* doc(templates): update user story for consistency

* chore(workflows): move GitHub Actions release configuration files to keep within corresponding domain

* chore(devcontainer): move compose file to its corresponding domain

* chore(devcontainer): add configuration file to exclude files from being copied on building process

* doc(README): remove files that will be created later with updated version

* chore(service): remove compose file for cleanup

* doc(README): remove sections that will be updated

* chore(service): restructure project files for better maintainability

* refactor(service): update imports order for easy maintainability

* refactor(service): update imports to work with new project structure

* chore(service): update imports and configuration files to reflext new project structure changes

* chore(service): update Dockerfile to remove honcho prefix from logs for better readability

* chore(devcontainer): update configuration files to solve ownerhsip issues with project root directory
