+++ .devcontainer/.dockerignore
@@ -1,0 +1,6 @@
+devcontainer.json
+Dockerfile
+devcontainer.json
+Dockerfile
+docker-compose-development.yml
+.dockerignore
