+++ .github/ISSUE_TEMPLATE/user-story.md
@@ -14,9 +14,8 @@ assignees: yifattih
 ## Details and Assumptions
 - document what you know
 
-# Acceptance Criteria
+## Acceptance Criteria
 ```
-gherkin
 Given [some context]
 When [certain action is taken]
 Then [the outcome of action is observed]
