+++ .gitignore
@@ -1,4 +1,5 @@
 # Byte-compiled / optimized / DLL files
+**/__pycache__/
 __pycache__/
 *.py[cod]
 *$py.class
