--- .releaserc.yml
@@ -1,9 +1,0 @@
-plugins:
-  - "@semantic-release/commit-analyzer"
-  - "@semantic-release/release-notes-generator"
-  - "@semantic-release/changelog"
-  - "@semantic-release/github"
-
-branches:
-  - "+([0-9])?(.{+([0-9]),x}).x"
-  - main
