--- CODE_OF_CONDUCT.md
@@ -1,55 +1,0 @@
-# Contributor Code of Conduct
-
-## Our Pledge
-We as members, contributors, and leaders pledge to make participation in our project and our community a harassment-free experience for everyone, regardless of age, body size, visible or invisible disability, ethnicity, sex characteristics, gender identity and expression, level of experience, education, socio-economic status, nationality, personal appearance, race, religion, or sexual identity and orientation.
-
-We pledge to act and interact in ways that contribute to an open, welcoming, diverse, inclusive, and healthy community.
-
----
-
-## Our Standards
-Examples of behavior that contribute to a positive environment include:
-- Demonstrating empathy and kindness towards other people
-- Being respectful of differing opinions, viewpoints, and experiences
-- Giving and gracefully accepting constructive feedback
-- Focusing on what is best for the community
-- Showing respect for other community members' time and effort
-
-Examples of unacceptable behavior include:
-- The use of sexualized language or imagery, and sexual attention or advances of any kind
-- Trolling, insulting, or derogatory comments, and personal or political attacks
-- Public or private harassment
-- Publishing others' private information, such as a physical or email address, without their explicit permission
-- Other conduct which could reasonably be considered inappropriate in a professional setting
-
----
-
-## Responsibilities
-Project maintainers are responsible for clarifying and enforcing standards of acceptable behavior and will take appropriate and fair corrective action in response to any behavior deemed inappropriate, threatening, offensive, or harmful.
-
-Project maintainers have the right and responsibility to remove, edit, or reject comments, commits, code, wiki edits, issues, and other contributions that are not aligned with this Code of Conduct.
-
----
-
-## Scope
-This Code of Conduct applies within all project spaces, including but not limited to GitHub repositories, pull requests, issue tracking, communication channels, and any official project-related events.
-
-It also applies when an individual is officially representing the project in public spaces.
-
----
-
-## Enforcement
-Instances of abusive, harassing, or otherwise unacceptable behavior may be reported to the project maintainers at **[project-maintainer@example.com](mailto:project-maintainer@example.com)**.
-
-All complaints will be reviewed and investigated promptly and fairly.
-
-Project maintainers are obligated to maintain confidentiality regarding the reporter of an incident.
-
-Further details of specific enforcement policies may be posted separately.
-
----
-
-## Attribution
-This Code of Conduct is adapted from the [Contributor Covenant](https://www.contributor-covenant.org/), version 2.1, available at [https://www.contributor-covenant.org/version/2/1/code_of_conduct/](https://www.contributor-covenant.org/version/2/1/code_of_conduct/).
-
-For answers to common questions about this code of conduct, see [https://www.contributor-covenant.org/faq](https://www.contributor-covenant.org/faq).
