--- docker-compose-devcontainer.yml
@@ -1,20 +1,0 @@
-services:
-  devcontainer:
-    container_name: development-env
-    build: 
-      context: .devcontainer
-      dockerfile: Dockerfile
-    volumes:
-      - ../..:/workspaces:cached      
-    command: sleep infinity
-    ports:
-      - "8000:8000"
-
-  service:
-    container_name: service
-    build:
-      context: service
-      dockerfile: Dockerfile
-    network_mode: service:devcontainer   # Puts app and devcontainer in the same network so browsing http://127.0.0.1:8080 works
-    environment:
-      - PORT=8000
