--- docker-compose-observability-stack.yml
@@ -1,29 +1,0 @@
-services:
-  service:
-    container_name: service
-    build:
-      context: ./service
-      dockerfile: Dockerfile
-    ports:
-      - "8001:8000"
-    depends_on:
-      - prometheus
-
-  prometheus:
-    container_name: prometheus
-    build:
-      context: ./prometheus
-      dockerfile: Dockerfile
-    ports:
-      - "9090:9090"
-
-  grafana:
-    container_name: grafana
-    image: grafana/grafana
-    ports:
-      - "3000:3000"
-    volumes:
-      - grafana-data:/var/lib/grafana
-
-volumes:
-  grafana-data:
