+++ service/Dockerfile
@@ -36,8 +36,8 @@ COPY --from=builder /root/.local /home/yifattih/.local
 ENV PATH="/home/yifattih/.local/bin:$PATH"
 
 # Copy only necessary application files
-COPY src/ /service/src/
-COPY Procfile /service/
+COPY Procfile main.py /service/
+COPY models /service/models
 
 # Add variable for Procfile execution via honcho
 ENV ENV="prod"
@@ -46,4 +46,4 @@ ENV ENV="prod"
 EXPOSE 8000
 
 # Start the application
-CMD ["honcho", "start"]
+CMD ["honcho", "start", "--no-prefix"]
