+++ service/main.py
@@ -1,11 +1,10 @@
 from datetime import datetime, timezone
 
 from fastapi import FastAPI, HTTPException, Response, status
+from models.metadata import Metadata
+from models.model import RMRModel
+from models.schemas import HealthStatus, InputData, OutputData
 from prometheus_fastapi_instrumentator import Instrumentator
-
-from .metadata import Metadata
-from .models.model import RMRModel
-from .schemas import HealthStatus, InputData, OutputData
 
 rmr_model = RMRModel()
 start_time_utc = datetime.now(timezone.utc)  # Store the time the api starts
